+++ beam-core/src/main/java/org/esa/beam/examples/codesnippets/framework/datamodel/ProductExamples.java
@@ -115,7 +115,7 @@ public class ProductExamples {
     public Product addTiePointGeoCoding(Product product) {
         // define the geolocation as two array containing latitude and longitude of the product for
         // a subsampled range of the pixels.
-        double[] latitudes = {
+        float[] latitudes = {
                 64.508275f,
                 64.673630f,
                 64.836693f,
@@ -126,7 +126,7 @@ public class ProductExamples {
                 64.101139f,
                 64.260961f
         };
-        double[] longitudes = {
+        float[] longitudes = {
                 39.620031f,
                 39.041526f,
                 38.455981f,
